requirements.txt
colorlog==6.7.0
discord.py==2.0.1
PyNaCl==1.5.0